api/requirements.txt
Django
django-extensions
django-ninja
django-ninja-apikey
django-ninja-crud
django-ninja-extra
django-ninja-jwt
django-split-settings
django-tenants
django-unfold
psycopg2
tox
libmagic
django-versatileimagefield
celery
django-celery-beat
django-ckeditor
django-admin-anchors
django-cors-headers
django-countries
django-environ
django-filer
django-import-export
django-mdeditor
django-nested-admin
django-o365mail
django-taggit
django-timezone-field
django-trench
djangorestframework
djangorestframework-simplejwt
gunicorn
flower
honcho
Markdown
markdownify
MarkupPy
MarkupSafe
O365
Pillow
model-bakery
faker
git+https://github.com/scalar/scalar.git#subdirectory=integrations/django-ninja
soundfile